+++ packages/frontend/synthesizer/core/src/synthesizer/types/buffers.ts
@@ -1,7 +1,7 @@
 import { DEFAULT_SOURCE_BIT_SIZE } from '../params/constants.ts';
-import type { DataPt, DataPtDescription } from './index.ts'
+import type { DataPtDescription } from './index.ts'
 import { FUNCTION_INPUT_LENGTH } from 'tokamak-l2js';
-import { BUFFER_LIST, ReservedBuffer, SubcircuitNames } from '../../subcircuit/configuredTypes.ts';
+import { BUFFER_LIST, ReservedBuffer } from '../../subcircuit/configuredTypes.ts';
 
 const PUBLIC_OUT_VARIABLES_STATIC = [
   // Nothing